Swithing m_Size to float

+++ assets/Materials/Geom/Hexamap/SimpleGrid.geom
@@ -5,7 +5,7 @@ layout (max_vertices = 7) out;
 uniform mat4 g_WorldViewProjectionMatrix;
 const float PI = 3.141592654;
 
-uniform int m_Size;
+uniform float m_Size;
 uniform bool m_Pointy;
 
 void main(){
